moved getapiprovider->provider pkg
